Update gcc version in build_toolchain.sh
I encountered this error (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86724) using gcc 8.2.0 and had to manually add in the include statements in 'graphite.h' in order to successfully build the toolchain. Later versions of gcc fix this issue, so this error hopefully won't crop up again for other users. I also updated the versioning of the other dependencies, but feel free to reject this if it breaks backward compatibility somewhere.