Do lcd_compat differently, using LCD SPI commands
The LCD panel controller (ILI9341?) supports rotation and mirroring in
hardware, which allows a configuration very close to the CX pre-W native
320x240 LCD. For working timing, it has to be treated as 320x320 though,
so remapping the LCD and adding abort handler for protecting the timing
registers is still necessary.

Generally the timing is only set up on program start though, so the abort
handler can be disabled after that, which means it's close to zero overhead
from that point on.

Unfortunately reading some registers over SPI is for some reason broken,
so the values for restoring the original state have to be hardcoded.