Reintroduce a timeout for the "Ndless installed!" message
This message is wrongly checked for in exams. This is worrisome, because
some patched Ndless to not show the message, so it's not even a reliable
indicator for Ndless, let alone the presence of cheating software.

This change should make this more obvious and hopefully lead to better
measures, like inspecting the documents folder or using exam mode.