Scan for an embedded Zehn file before loading the PRG binary into memory
Previously it allocated memory for the entire PRG binary, then found the
Zehn signature and called zehn_load while still holding on to that allocation.
That meant that the whole file had to fit into memory twice during loading,
or even three times in the case of compressed Zehn files.

This was by accident worked around in the case that emu_debug_alloc_ptr was
set: The PRG binary was loaded into that block but zehn_load also wrote into
that. With the preceding refactoring of emu_debug_alloc_ptr which included
enforcing that there's only one active user, this didn't work anymore either.

Now only the area needed for scanning is loaded into memory before calling
zehn_load or performing the full PRG binary load.